Cleanup: Removed lock files, added utils file for reusable code, and added a .env template.
