+++ .env_template
@@ -1,0 +1,4 @@
+ACCES_TOKEN=
+PASSWORD=
+EMAIL=
+UUID_SECRET=
