+++ backend/index.js
@@ -11,6 +11,7 @@ const jwt = require('./components/auth/jwt');
 const sendEmail = require('./components/email/sendEmail')
 const uuid = require('./components/auth/uuid');
 const createuser = require('./components/auth/createuser');
+const utils = require('./components/utils/utils.js');
 
 app.use(express.json());
 
@@ -52,6 +53,7 @@ app.post('/api/login', (req, res) => {
   const password = req.body.password;
 
   jwt.authenticateUser({ email, password }).then(user => {
+    console.log(user);
     res.json({ accessToken: accessToken, firstname: firstname });
   }).catch(err => {
     console.log(err)
@@ -66,7 +68,7 @@ app.post('/api/login', (req, res) => {
 app.post('/api/forgotpassword', (req, res) => {
   const email = req.body.email;
   //getuid
-  sendEmail.getUID({ email }).then(user => {
+  utils.getUID({ email }).then(user => {
     const mailOptions = {
       from: 'joshuajy03@gmail.com',
       to: `${email}`,
@@ -93,14 +95,16 @@ app.post('/api/resetpassword', (req, res) => {
     res.send('Password Updated')
   ).catch(err => {
     console.log(err)
-    
+  })
+})
+
 app.post('/api/createuser', (req, res) => {
 
   createuser.createuser(req.body).then((user) => {
     res.json({"user_id": user.user_id, email: req.body.email, firstname: req.body.firstname});
   }).catch(err => {
     console.log(JSON.stringify(err))
-    res.sendStatus(err.error);
+    res.sendStatus(500);
   });
 
 })
@@ -110,8 +114,8 @@ function authenticateToken(req, res, next) {
   if (token == null) {
     // we don't have a token
     res.sendStatus(401);
-  });
-});
+  };
+};
 
 app.listen(port, () => {
   console.log(`BoilerTime API listening on port ${port}!`)
