+++ backend/package.json
@@ -26,6 +26,7 @@
     "express": "^4.18.2",
     "firebase": "^9.17.1",
     "firebase-admin": "^11.5.0",
-    "nodemailer": "^6.9.1"
+    "nodemailer": "^6.9.1",
+    "nodemon": "^2.0.20"
   }
 }
